requirements.txt
about-time==4.2.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alibabacloud-actiontrail20200706==2.4.1
alibabacloud-credentials==1.0.3
alibabacloud-credentials-api==1.0.0
alibabacloud-cs20151215==6.1.0
alibabacloud-darabonba-string==0.0.4
alibabacloud-ecs20140526==7.2.5
alibabacloud-gateway-sls==0.4.0
alibabacloud-gateway-sls-util==0.4.0
alibabacloud-oss20190517==1.0.6
alibabacloud-ram20150501==1.2.0
alibabacloud-rds20140815==12.0.0
alibabacloud-sas20181203==6.1.0
alibabacloud-sls20201230==5.9.0
alibabacloud-sts20150401==1.1.6
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.1
alibabacloud-tea-util==0.3.14
alibabacloud-vpc20160428==6.13.0
alibabacloud_darabonba_array==0.1.0
alibabacloud_darabonba_encode_util==0.0.2
alibabacloud_darabonba_map==0.0.1
alibabacloud_darabonba_signature_util==0.0.4
alibabacloud_darabonba_time==0.0.1
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_oss==0.0.17
alibabacloud_gateway_oss_util==0.0.3
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_oss_util==0.0.6
alibabacloud_tea_xml==0.0.3
alive-progress==3.3.0
aliyun-log-fastpb==0.2.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
awsipranges==0.3.3
azure-common==1.1.28
azure-core==1.37.0
azure-identity==1.21.0
azure-keyvault-keys==4.10.0
azure-mgmt-apimanagement==5.0.0
azure-mgmt-applicationinsights==4.1.0
azure-mgmt-authorization==4.0.0
azure-mgmt-compute==34.0.0
azure-mgmt-containerregistry==12.0.0
azure-mgmt-containerservice==34.1.0
azure-mgmt-core==1.6.0
azure-mgmt-cosmosdb==9.7.0
azure-mgmt-databricks==2.0.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-loganalytics==12.0.0
azure-mgmt-monitor==6.0.2
azure-mgmt-network==28.1.0
azure-mgmt-postgresqlflexibleservers==1.1.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-recoveryservices==3.1.0
azure-mgmt-recoveryservicesbackup==9.2.0
azure-mgmt-resource==23.3.0
azure-mgmt-search==9.1.0
azure-mgmt-security==7.0.0
azure-mgmt-sql==3.0.1
azure-mgmt-storage==22.1.1
azure-mgmt-subscription==3.1.1
azure-mgmt-web==8.0.0
azure-monitor-query==2.0.0
azure-storage-blob==12.24.1
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.39.15
botocore==1.39.15
build==1.3.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.0
ci-info==0.3.0
circuitbreaker==2.1.3
click==8.3.1
click-plugins==1.1.1.2
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
contextlib2==21.6.0
contourpy==1.3.3
cryptography==44.0.1
cycler==0.12.1
darabonba-core==1.0.5
dash==3.1.1
dash-bootstrap-components==2.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
detect-secrets==1.5.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.8.0
dulwich==0.23.0
durationpy==0.10
email_validator==2.2.0
etelemetry==0.3.1
filelock==3.20.2
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
fpdf==1.7.2
frontend==0.0.3
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.28.1
google-api-python-client==2.163.0
google-auth==2.47.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
graphemeu==0.7.2
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
iamdata==0.1.202601071
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==32.0.1
langsmith==0.6.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.12
llama-index-cli==0.5.3
llama-index-core==0.14.12
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.12
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-chroma==0.5.5
llama-index-workflows==2.11.7
llama-parse==0.6.54
looseversion==1.3.0
lxml==6.0.2
lz4==4.4.5
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
microsoft-kiota-abstractions==1.9.2
microsoft-kiota-authentication-azure==1.9.2
microsoft-kiota-http==1.9.2
microsoft-kiota-serialization-form==1.9.2
microsoft-kiota-serialization-json==1.9.2
microsoft-kiota-serialization-multipart==1.9.2
microsoft-kiota-serialization-text==1.9.2
mmh3==5.2.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgraph-core==1.3.8
msgraph-sdk==1.23.0
msrest==0.7.1
multidict==6.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
nltk==3.9.2
numpy==2.0.2
nvidia-cusparselt-cu12==0.7.1
oauthlib==3.3.1
oci==2.160.3
onnxruntime==1.23.2
openai==2.14.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
prov==2.1.1
prowler==5.16.1
psycopg2-binary==2.9.11
puremagic==1.30
py-iam-expand==0.1.0
py-ocsf-models==0.5.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
pyfair==0.1a13
PyGithub==2.5.0
Pygments==2.19.2
PyJWT==2.10.1
pymssql==2.3.11
PyMuPDF==1.26.7
PyNaCl==1.6.2
pyodbc==5.3.0
pyOpenSSL==24.3.0
pyparsing==3.3.1
pypdf==6.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.5.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.13.1
schema==0.7.5
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
shodan==1.31.0
simplejson==3.20.2
six==1.17.0
slack_sdk==3.34.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
std-uritemplate==2.0.8
streamlit==1.52.2
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tldextract==5.3.1
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
traits==7.1.0
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
zstd==1.5.7.2
